+++ odoo/local-src/sf_stock/models/stock_move.py
@@ -10,6 +10,27 @@ class StockMove(models.Model):
 
     _inherit = "stock.move"
 
+    def _compute_amount(self):
+        """Compute the stock move price value
+        based on the information of the related sale line"""
+        for move in self:
+            sale_line_id = move.procurement_id.sale_line_id
+            if sale_line_id:
+                price = sale_line_id.price_unit * \
+                        (1 - (sale_line_id.discount or 0.0) / 100.0)
+                taxes = sale_line_id.tax_id.compute_all(
+                    price,
+                    sale_line_id.currency_id,
+                    move.product_uom_qty,
+                    product=move.product_id,
+                    partner=sale_line_id.order_id.partner_shipping_id)
+                move.update({
+                    'price_tax': taxes['total_included'] - taxes[
+                        'total_excluded'],
+                    'price_total': taxes['total_included'],
+                    'price_subtotal': taxes['total_excluded'],
+                })
+
     @api.multi
     def action_done(self):
         super(StockMove, self).action_done()
@@ -43,3 +64,20 @@ class StockMove(models.Model):
                         'first_outgoing_stock_move_id': move.id
                     })
         return True
+
+    currency_id = fields.Many2one(
+        related='procurement_id.sale_line_id.currency_id',
+        string='Currency',
+        readonly=True)
+    price_subtotal = fields.Monetary(
+        compute='_compute_amount',
+        string='Subtotal', readonly=True
+    )
+    price_tax = fields.Monetary(
+        compute='_compute_amount',
+        string='Taxes', readonly=True
+    )
+    price_total = fields.Monetary(
+        compute='_compute_amount',
+        string='Total', readonly=True
+    )