+++ odoo/songs/install/data_demo.py
@@ -48,12 +48,6 @@ def import_crm_team(ctx):
 
 
 @anthem.log
-def import_warehouse(ctx):
-    """ Importing stock warehouse from csv """
-    load_csv(ctx, 'data/demo/stock_warehouse.csv', 'stock.warehouse')
-
-
-@anthem.log
 def import_location(ctx):
     """ Importing stock location from csv """
     load_csv(ctx, 'data/demo/stock_location.csv', 'stock.location')
@@ -111,7 +105,6 @@ def main(ctx):
     import_customers(ctx)
     import_suppliers(ctx)
     import_crm_team(ctx)
-    import_warehouse(ctx)
     import_location(ctx)
     import_customers(ctx)
     import_drone_type(ctx)